requirements.txt
django
psycopg2
pyxero
cryptography
